forecast-critic/pyproject.toml
[project]
name = "forecast-critic"
version = "0.1.0"
description = "The Forecast Critic: LLM-based automated forecast monitoring"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "matplotlib>=3.8",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "pandas>=2.1",
    "tqdm>=4.66",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
gemini = ["google-genai>=1.0"]
m5 = [
    "torch>=2.1",
    "chronos-forecasting>=1.0",
]

[project.scripts]
forecast-critic = "main:main"
